+++ src/main/java/eci/server/DashBoardModule/dto/ProjectDashboardDto.java
@@ -1,15 +1,13 @@
-package eci.server.ProjectModule.dto;
+package eci.server.DashBoardModule.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import eci.server.ItemModule.dto.item.ItemProjectDashboardDto;
-import eci.server.ItemModule.dto.item.ItemProjectDto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.util.List;
 
 @Data
 @AllArgsConstructor
@@ -38,6 +36,5 @@ public class ProjectDashboardDto {
 
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss", timezone = "Asia/Seoul")
     private LocalDateTime createdAt;
-
 
 }
