+++ src/main/java/eci/server/ProjectModule/controller/page/ProjectPageController.java
@@ -1,22 +1,11 @@
-package eci.server.ItemModule.controller.page;
+package eci.server.ProjectModule.controller.page;
 
-import eci.server.ItemModule.dto.item.ItemProjectDashboardDto;
 import eci.server.ItemModule.dto.item.ItemProjectDto;
-import eci.server.ItemModule.dto.item.ItemSimpleDto;
-import eci.server.ItemModule.entity.item.Item;
-import eci.server.ItemModule.exception.member.MemberNotFoundException;
-import eci.server.ItemModule.repository.item.ItemRepository;
-import eci.server.ItemModule.repository.member.MemberRepository;
-import eci.server.ItemModule.repository.newRoute.RouteOrderingRepository;
-import eci.server.ItemModule.repository.newRoute.RouteProductRepository;
-import eci.server.ProjectModule.dto.ProjectDashboardDto;
 import eci.server.ProjectModule.dto.project.ProjectMemberRequest;
 import eci.server.ProjectModule.dto.project.ProjectSimpleDto;
 import eci.server.ProjectModule.entity.project.Project;
 import eci.server.ProjectModule.entity.projectAttachment.ProjectAttachment;
-import eci.server.ProjectModule.exception.ProjectNotFoundException;
 import eci.server.ProjectModule.repository.project.ProjectRepository;
-import eci.server.ProjectModule.service.ProjectService;
 import eci.server.aop.AssignMemberId;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -30,52 +19,11 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import javax.transaction.Transactional;
-import javax.validation.constraints.Null;
 import java.util.stream.Collectors;
 @Transactional
 @RestController
 @RequiredArgsConstructor
-public class PageController {
-
-    @Autowired
-    ItemRepository itemRepository;
-
-    @CrossOrigin(origins = "https://localhost:3000")
-    @GetMapping("/item/page")
-    public Page<ItemSimpleDto> paging(@PageableDefault(size=5)
-                                      @SortDefault.SortDefaults({
-                                              @SortDefault(
-                                                      sort = "createdAt",
-                                                      direction = Sort.Direction.DESC)
-                                      })
-                                      Pageable pageRequest) {
-
-        Page<Item> itemList = itemRepository.findAll(pageRequest);
-
-        Page<ItemSimpleDto> pagingList = itemList.map(
-                item -> new ItemSimpleDto(
-
-                        item.getId(),
-                        item.getName(),
-                        item.getType(),
-                        item.getItemNumber(),
-                        item.getWidth(),
-                        item.getHeight(),
-                        item.getWeight(),
-                        item.getMember().getUsername(),
-                        item.getMaterials().get(0).getMaterial().getName(),
-                        item.getColor().getColor(),
-                        item.getThumbnail().get(0).getImageaddress(),
-                        item.getAttachments().stream().map(
-                                a -> a.getAttachmentaddress()
-                        ).collect(Collectors.toList()),
-                        item.getCreatedAt()
-
-                )
-        );
-
-        return pagingList;
-    }
+public class ProjectPageController {
 
     /**
      * 프로젝트 모듈에서의 프로젝트 리스트 (내가 만든 프로젝트들 maybe..?)
@@ -139,29 +87,5 @@ public class PageController {
 
         return pagingList;
     }
-    //이걸로 프로젝트의 라우트 오더링, 프로덕트에 해당하는 아이들을 찾아서 데려올 것이다
-    @Autowired
-    ProjectService projectService;
-    @CrossOrigin(origins = "https://localhost:3000")
-    @GetMapping("dashboard/project/page")
-    @AssignMemberId //작성한 멤버
-    public Page<ProjectDashboardDto> pagingDashboardProject(@PageableDefault(size=5)
-                                                            @SortDefault.SortDefaults({
-                                                                    @SortDefault(
-                                                                            sort = "createdAt",
-                                                                            direction = Sort.Direction.DESC
-                                                                    )
-                                                            }
-                                                            )
-                                                                    Pageable pageRequest,
-                                                            ProjectMemberRequest req) {
 
-        Page<ProjectDashboardDto> projectDashboardDtos = projectService.readDashboard(
-                pageRequest, req
-        );
-
-
-        return projectDashboardDtos;
-
-    }
 }
