+++ src/main/java/eci/server/ProjectModule/dto/project/ReadProjectDto.java
@@ -1,4 +1,4 @@
-package eci.server.ProjectModule.dto;
+package eci.server.ProjectModule.dto.project;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import eci.server.ItemModule.dto.color.ColorDto;
