+++ src/main/java/eci/server/ProjectModule/dto/projectTodo/ProjectTodoResponse.java
@@ -1,0 +1,15 @@
+package eci.server.ProjectModule.dto.projectTodo;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class ProjectTodoResponse {
+    Long projectId;
+    String projectName;
+    String projectType;
+    Integer projectNumber;
+}
